refactor: qna Posts
