+++ package.json
@@ -30,6 +30,7 @@
     "react-intersection-observer": "^9.13.0",
     "react-spinners": "^0.14.1",
     "react-toastify": "^10.0.5",
+    "sharp": "^0.33.4",
     "swiper": "^11.1.9",
     "tailwindcss-animate": "^1.0.7",
     "uuidv4": "^6.2.13",
