--- src/app/(home)/(posts)/_components/qna/PopularQnaPosts.tsx
@@ -1,103 +1,0 @@
-'use client';
-
-import GradCap from '@/assets/images/qna/GradCap';
-import { cutText, processMarkdown } from '@/utils/markdownCut';
-import useFetchQnaPosts from '@/hooks/qna/useFetchQnaPosts';
-import MDEditor from '@uiw/react-md-editor';
-import Link from 'next/link';
-
-const Loading = () => <div>Loading...</div>;
-
-const Error = ({ message }: { message: string }) => <div>Error: {message}</div>;
-
-const PopularQnaPosts = () => {
-  const {
-    data: popularPosts,
-    error: popularError,
-    isPending: isPendingPopular,
-    isError: isErrorPopular,
-    page: popularPage,
-    totalPages: popularTotalPages,
-    goToPage: goToPopularPage
-  } = useFetchQnaPosts('popular');
-
-  if (isPendingPopular) {
-    return <Loading />;
-  }
-
-  if (isErrorPopular) {
-    return <Error message={popularError?.message || 'Unknown error occurred'} />;
-  }
-
-  const pageSize = 6;
-  const startIndex = popularPage * pageSize;
-
-  const buttonClass = (index: number) => {
-    return `w-[33px] h-[32px] flex justify-center items-center flex-grow-0 flex-shrink-0 relative gap-2 px-3 py-1 rounded-md ${
-      index === popularPage ? 'bg-main-50 text-main-500' : 'bg-neutral-100 text-neutral-500 border border-neutral-100'
-    }`;
-  };
-
-  return (
-    <div className="w-[1204px] mx-auto p-4">
-      <div className="flex justify-start items-center relative gap-1.5 mb-8">
-        <p className="flex items-center text-h4 font-bold text-left text-neutral-900">
-          인기 QnA
-          <div className="ml-1">
-            <GradCap />
-          </div>
-        </p>
-      </div>
-      {popularPosts && popularPosts.length > 0 ? (
-        <ul className="grid grid-cols-2">
-          {popularPosts.map((post, index) => (
-            <Link key={post.id} href={`/qna/${post.id}`}>
-              <li
-                className={`border-t border-b border-neutral-100 p-4 flex flex-col min-h-[158px] ${
-                  index % 2 === 0 ? 'border-r' : 'border-l'
-                }`}
-              >
-                <div className="flex justify-start items-center mb-5">
-                  <p className="text-xl font-bold text-left text-main-500">{startIndex + index + 1}</p>
-                  <p className="ml-2 text-xl font-bold text-left text-neutral-900">{cutText(post.title, 40)}</p>
-                </div>
-                <div className="text-lg font-medium text-left text-neutral-300 mb-5" data-color-mode="light">
-                  <MDEditor.Markdown source={processMarkdown(post.content, 30)} />
-                </div>
-                <div className="mt-auto">
-                  <div className="flex justify-start items-center gap-2">
-                    <div className="flex items-center gap-1">
-                      <p className="text-body1 text-left text-main-500">좋아요</p>
-                      <p className="text-body1 text-left text-main-500">{post.qna_like?.[0]?.count || 0}</p>
-                    </div>
-                    <div className="w-0.5 h-[18px] bg-neutral-200" />
-                    <div className="flex items-center gap-1">
-                      <p className="text-lg text-left text-neutral-300">답변수</p>
-                      <p className="text-lg text-left text-neutral-300">{post.qna_comment?.[0]?.count || 0}</p>
-                    </div>
-                  </div>
-                </div>
-              </li>
-            </Link>
-          ))}
-        </ul>
-      ) : (
-        <div>게시물이 없습니다.</div>
-      )}
-      <div className="flex justify-center items-center gap-4 mt-8">
-        {Array.from({ length: popularTotalPages }, (_, index) => (
-          <button
-            key={index}
-            onClick={() => goToPopularPage(index)}
-            disabled={index === popularPage}
-            className={buttonClass(index)}
-          >
-            <p className="flex-grow-0 flex-shrink-0 text-body1 font-medium text-left">{index + 1}</p>
-          </button>
-        ))}
-      </div>
-    </div>
-  );
-};
-
-export default PopularQnaPosts;
